+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/subs.cljs
@@ -4,6 +4,7 @@
     [re-frame.core :refer [reg-sub subscribe]]
     [sixsq.nuvla.ui.credentials.subs :as creds-subs]
     [sixsq.nuvla.ui.deployment-dialog.spec :as spec]
+    [sixsq.nuvla.ui.deployment-dialog.utils :as utils]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]))
 
 
@@ -86,16 +87,36 @@
 
 
 (reg-sub
+  ::execution-mode
+  :<- [::deployment]
+  :<- [::selected-infra-service]
+  :<- [::selected-credential-id]
+  (fn [[deployment infra-service cred-id]]
+    (or (:execution-mode deployment)
+        (let [cred-unknown? @(subscribe [::creds-subs/credential-check-status-unknown? cred-id])
+              cred-loading? @(subscribe [::creds-subs/credential-check-loading? cred-id])]
+          (if (utils/infra-support-pull? infra-service)
+            (if (and cred-unknown? (not cred-loading?))
+              "pull"
+              "mixed")
+            "push")))))
+
+
+(reg-sub
   ::modal-action-button-text
   :<- [::i18n-subs/tr]
   :<- [::deployment-start?]
   :<- [::is-launch-status? :ok]
-  (fn [[tr start? launch-status-ok?]]
-    (tr [(cond
-           (and start? launch-status-ok?) :launch
-           (and start? (not launch-status-ok?)) :launch-force
-           (and (not start?) launch-status-ok?) :update
-           :else :update-force)])))
+  :<- [::execution-mode]
+  (fn [[tr start? launch-status-ok? execution-mode]]
+    (let [execution-mode-pull? (= execution-mode "pull")]
+      (tr [(cond
+             (and start? execution-mode-pull?) :schedule-launch
+             (and (not start?) execution-mode-pull?) :schedule-update
+             (and start? launch-status-ok?) :launch
+             (and start? (not launch-status-ok?)) :launch-force
+             (and (not start?) launch-status-ok?) :update
+             :else :update-force)]))))
 
 
 (reg-sub
@@ -120,12 +141,12 @@
         registries-completed? credentials-completed? registries-status]
        [_ step-id]]
     (let [cred-loading? @(subscribe [::creds-subs/credential-check-loading? cred-id])
-          cred-invalid? @(subscribe [::creds-subs/credential-check-status-invalid? cred-id])]
+          cred-valid?   @(subscribe [::creds-subs/credential-check-status-valid? cred-id])]
       (case step-id
         :data data-completed?
         :infra-services (and credentials-completed?
                              (not cred-loading?)
-                             (not cred-invalid?))
+                             cred-valid?)
         :env-variables env-variables-completed?
         :registries (and registries-completed? (= :ok registries-status))
         :license license-completed?
@@ -558,9 +579,12 @@
   :<- [::price]
   :<- [::price-completed?]
   :<- [::version-completed?]
+  :<- [::selected-credential-id]
   (fn [[deployment data-completed? data-step-active? credentials-completed? env-variables-completed?
-        registries-completed? license license-completed? price price-completed? version-completed?]]
-    (or (not deployment)
+        registries-completed? license license-completed? price price-completed? version-completed?
+        selected-credential-id]]
+    (let [cred-invalid? @(subscribe [::creds-subs/credential-check-status-invalid? selected-credential-id])]
+      (or (not deployment)
         (and (not data-completed?) data-step-active?)
         (not credentials-completed?)
         (not env-variables-completed?)
@@ -568,7 +592,8 @@
         (and price (not price-completed?))
         (and license (not license-completed?))
         (not registries-completed?)
-        (not version-completed?))))
+        (not version-completed?)
+        cred-invalid?))))
 
 
 (reg-sub